requirements.txt
Django<1.9
djangorestframework==3.1.3
psycopg2==2.5.1
